Search Highlighting
Added functionality to control the background in addition to the foreground text color.
This was done by adding the background color on the end of the foreground color already specified (e.g. 39 -> 3949).
Wrote a function to return either the first two numbers or last two numbers using < 1 or >=1. Used function wherever color was defined (besides defaulting at the end of a line; TODO).
Defined syntax highlighting colors as constants, pulling them out of the switch and putting them at the top of the program.
Eventually I would like to have these more easily editable through a CUI of some sort; TODO.